updateStandings takes finishedParticipants into account
