+++ src/main/java/org/braekpo1nt/mctmanager/games/game/footrace/states/ActiveState.java
@@ -128,6 +128,7 @@ public class ActiveState implements FootRaceState {
                                 context.getFinishedParticipants().indexOf(playerUUID) + 1))
                         .append(Component.text("!")))
         );
+        context.updateStandings();
     }
     
     @Override
